rename
# Conflicts:
#	static/app/views/performance/onboarding.spec.tsx
