+++ static/app/components/modals/redirectToProject.spec.tsx
@@ -3,7 +3,7 @@ import {act, renderGlobalModal} from 'sentry-test/reactTestingLibrary';
 
 import {openModal} from 'sentry/actionCreators/modal';
 import {RedirectToProjectModal} from 'sentry/components/modals/redirectToProject';
-import {testableWindowLocation} from 'sentry/utils/testableLocation';
+import {testableWindowLocation} from 'sentry/utils/testableWindowLocation';
 
 jest.unmock('sentry/utils/recreateRoute');
