+++ static/app/utils/testableWindowLocation.tsx
